Move Config Attribute info into config.rs
