+++ core/src/ast/attrs.rs
@@ -1,6 +1,5 @@
 //! This module contains utilities for dealing with Rust attributes
 
-use quote::ToTokens;
 use serde::ser::{SerializeStruct, Serializer};
 use serde::Serialize;
 use std::borrow::Cow;
@@ -39,9 +38,6 @@ pub struct Attrs {
     /// not see the impl blocks.
     pub attrs: Vec<DiplomatBackendAttr>,
 
-    /// Backend specific configuration attributes:
-    pub config_attrs: Vec<DiplomatBackendConfigAttr>,
-
     /// Renames to apply to the underlying C symbol. Can be found on methods, impls, and bridge modules, and is inherited.
     ///
     /// Affects method names when inherited onto methods.
@@ -60,7 +56,6 @@ impl Attrs {
         match attr {
             Attr::Cfg(attr) => self.cfg.push(attr),
             Attr::DiplomatBackend(attr) => self.attrs.push(attr),
-            Attr::DiplomatBackendConfig(attr) => self.config_attrs.push(attr),
             Attr::CRename(rename) => self.abi_rename.extend(&rename),
             Attr::DemoBackend(attr) => self.demo_attrs.push(attr),
         }
@@ -91,7 +86,6 @@ impl Attrs {
             cfg: self.cfg.clone(),
 
             attrs,
-            config_attrs: self.config_attrs.clone(),
             abi_rename,
             demo_attrs,
         }
@@ -118,7 +112,6 @@ impl From<&[Attribute]> for Attrs {
 enum Attr {
     Cfg(Attribute),
     DiplomatBackend(DiplomatBackendAttr),
-    DiplomatBackendConfig(DiplomatBackendConfigAttr),
     CRename(RenameAttr),
     DemoBackend(DemoBackendAttr),
     // More goes here
@@ -127,7 +120,6 @@ enum Attr {
 fn syn_attr_to_ast_attr(attrs: &[Attribute]) -> impl Iterator<Item = Attr> + '_ {
     let cfg_path: syn::Path = syn::parse_str("cfg").unwrap();
     let dattr_path: syn::Path = syn::parse_str("diplomat::attr").unwrap();
-    let dconfig_path: syn::Path = syn::parse_str("diplomat::config").unwrap();
     let crename_attr: syn::Path = syn::parse_str("diplomat::abi_rename").unwrap();
     let demo_path: syn::Path = syn::parse_str("diplomat::demo").unwrap();
     attrs.iter().filter_map(move |a| {
@@ -137,11 +129,6 @@ fn syn_attr_to_ast_attr(attrs: &[Attribute]) -> impl Iterator<Item = Attr> + '_ 
             Some(Attr::DiplomatBackend(
                 a.parse_args()
                     .expect("Failed to parse malformed diplomat::attr"),
-            ))
-        } else if a.path() == &dconfig_path {
-            Some(Attr::DiplomatBackendConfig(
-                a.parse_args()
-                    .expect("Failed to parse malformed diplomat::config"),
             ))
         } else if a.path() == &crename_attr {
             Some(Attr::CRename(RenameAttr::from_meta(&a.meta).unwrap()))
@@ -273,58 +260,6 @@ impl Parse for DiplomatBackendAttr {
         let _comma: Token![,] = input.parse()?;
         let meta = input.parse()?;
         Ok(Self { cfg, meta })
-    }
-}
-
-#[derive(Clone, PartialEq, Eq, Hash, Debug, Serialize)]
-#[non_exhaustive]
-pub struct DiplomatBackendConfigAttr {
-    pub key_value_pairs: Vec<DiplomatBackendConfigKeyValue>,
-}
-
-impl Parse for DiplomatBackendConfigAttr {
-    fn parse(input: ParseStream) -> syn::Result<Self> {
-        let list = input.parse_terminated(DiplomatBackendConfigKeyValue::parse, Token![,])?;
-        let vec = list.into_iter().collect();
-        Ok(Self {
-            key_value_pairs: vec,
-        })
-    }
-}
-
-#[derive(Clone, PartialEq, Eq, Hash, Debug, Serialize)]
-#[non_exhaustive]
-pub struct DiplomatBackendConfigKeyValue {
-    pub key: String,
-    pub value: String,
-}
-
-impl Parse for DiplomatBackendConfigKeyValue {
-    fn parse(input: ParseStream) -> syn::Result<Self> {
-        let mut key_str: Vec<String> = Vec::new();
-
-        loop {
-            let i: Ident = input.parse()?;
-
-            key_str.push(i.to_string());
-
-            if input.peek(Token![.]) {
-                let _period: Token![.] = input.parse()?;
-            } else {
-                break;
-            }
-        }
-
-        let _equals: Token![=] = input.parse()?;
-
-        let val_expr: Expr = input.parse()?;
-
-        let value = val_expr.to_token_stream().to_string();
-
-        Ok(Self {
-            key: key_str.join("."),
-            value,
-        })
     }
 }
 
@@ -522,7 +457,7 @@ mod tests {
     use syn;
 
     use super::{
-        DiplomatBackendAttr, DiplomatBackendAttrCfg, DiplomatBackendConfigAttr, RenameAttr,
+        DiplomatBackendAttr, DiplomatBackendAttrCfg, RenameAttr,
     };
 
     #[test]
@@ -545,13 +480,6 @@ mod tests {
         let attr: syn::Attribute =
             syn::parse_quote!(#[diplomat::attr(any(cpp, has = "overloading"), namespacing)]);
         let attr: DiplomatBackendAttr = attr.parse_args().unwrap();
-        insta::assert_yaml_snapshot!(attr);
-    }
-
-    #[test]
-    fn test_cfg_attr() {
-        let attr: syn::Attribute = syn::parse_quote!(#[diplomat::config(test.out = 24, other.out =test, somefinal.out= "testing spaces")]);
-        let attr: DiplomatBackendConfigAttr = attr.parse_args().unwrap();
         insta::assert_yaml_snapshot!(attr);
     }
